asoc: wsa883x: Enable/Disable swr ports based on setbit.
During some concurrencies even though we are not
enabling the swrm port, we are trying to disable it.
which causes problem w.r.t clock disablement,
To avoid that we are updating the set bit only
when port is enabled, based on that bit we are taking
decision to disable or enable the port.

Change-Id: Iac9769515e17a228e31eebb052d1efac313e679d
Signed-off-by: Ganapathiraju Sarath Varma <quic_ganavarm@quicinc.com>