asoc: codecs: Add logic to handle TX Dec Mute from Slave
In low-tier targets which do not have WSA Macro for VISense,
VI data is captured in the TX Macro and RX data is through
RX Macro.
Currently LPASS TX DEC is unmuted before the the PA/Codec
connected to RX Macro is turned on, hence when the MBAIS
and VI Buffers are turned on there is a glitch observed
at the start of VI Data.
Add logic to trigger TX DEC Mute and Unmute from slave to
LPASS TX to fix the issue.

Change-Id: Ief38a2b07974d23790d83c1367756807a4915c5e
Signed-off-by: Swapnil Kangralkar <quic_skangral@quicinc.com>