asoc: sdca: get the reg defaults and check for SDCA reg
In the regmap_config, retrieve the register defaults from the extension driver.Add logic to check for the SDCA register before updating the regmap_writeable and regmap_readable. The swr_connect_port and datapath should not be enabled before graph_open, as this is incorrect. Instead, the swr_connect_port and datapath should be invoked as part of pcm_prepare.

Change-Id: I3bf3a59c79b3977f88f1f3b667897df68f03075c
Signed-off-by: Prasad Kumpatla <quic_pkumpatl@quicinc.com>