asoc: codecs: mute the dec in hw_params dai_op
Currently, the DEC mute operation is performed before
the start of the use case in the enable_dec widget
callback, while the unmute operation is handled in
the mute_stream dai op. This can lead to inconsistent
behavior when DAIs are active while widgets are reset.

To ensure consistent DEC behavior, handle both mute
and unmute operations within the dai_ops. This change
will prevent unexpected behavior and ensure that
DEC remains in a predictable state.

Change-Id: Ie1bcb96feab2e4a4e164ffe608933e174f8440bd
Signed-off-by: Faiz Nabi Kuchay <quic_fkuchay@quicinc.com>