asoc: codecs: wsa885x_i2c: add virtual BE support for pandeiro i2c
for i2s dual mono function of pandeiro, we need to have 2 streams
active on the same backend, which is not possible in current
implementation.

Add Virtual Backend support for 2 streams to run on the same
backend, which makes it possible for haptics and playback to run
concurrently on the same Pandeiro instance.

Change-Id: I2ce53a74096791758a2895997887d8e10cabe500
Signed-off-by: Faiz Nabi Kuchay <fkuchay@qti.qualcomm.com>