ASoC: codecs: QMP: Fix issue with multiple use-cases
Fix issue with multiple use-cases running on the same dai_id.

In existing code for 2 use-cases having the same dai_id
the num_ch gets out of sync as we are not handling tear-down
for the 1st dai_id linked use-case which causes the driver to
think that 2nd use-case is just another channel for the first
use-case.

Improve the logging in aggregator.c file to make it easier
to debug.

Change-Id: I00f9810464a746be0f1fa9983d285a72b62c6f42
Signed-off-by: Faiz Nabi Kuchay <fkuchay@qti.qualcomm.com>