asoc: lpass-cdc: add dec_ref_count to avoid mute issue
In voice call scenarios, when switching to speaker-mic,
the UPD mic gets disabled after the speaker-mic is enabled.
As a result, one of the speaker-mic microphones becomes
muted because the UPD mic shares the same DEC
with the speaker-mic.To handle this, dec_ref_count is
added to check if any DEC is currently active.

Change-Id: I4483a6459ddba6eefaced560e6bfb7ad2298269e
Signed-off-by: Yuhui Zhao <yuhuzhao@qti.qualcomm.com>
(cherry picked from commit c3b2689a7575cfe032bebd05c1d8c33358fe4138)