ASoc: audio_mcahine: Error handling in wcd_used
There's inconsistent error handling in the wcd_used assignment.
When checking for WCD939X_DRV_NAME, if the component is not found,
it checks for WCD9378_DRV_NAME, and if that's not found, it checks for WCD937X_DRV_NAME.
However, the wcd_used assignment is done before the nested checks complete,
which could lead to incorrect codec identification.

Change-Id: I3f898b3d6ca9bcd577ac198234139a26b607a18a
Signed-off-by: Ravulapati Vishnu Vardhan Rao <visr@qti.qualcomm.com>